+++ src/variables.mjs
@@ -24,6 +24,7 @@ export { default as empty } from './variables/empty.mjs';
 export { default as get_debug_type } from './variables/get_debug_type.mjs';
 export { default as gettype } from './variables/gettype.mjs';
 export { default as to_int } from './variables/to_int.mjs';
+export { default as to_integer } from './variables/to_integer.mjs';
 export { default as isset } from './variables/isset.mjs';
 export { default as strval } from './variables/strval.mjs';
 export { default as print_r } from './variables/print_r.mjs';